+++ verify/linalg/tutte.test.cpp
@@ -11,7 +11,7 @@ using namespace cp_algo::math;
 using namespace cp_algo::linalg;
 using namespace cp_algo::random;
 
-const int mod = 998244353;
+const int64_t mod = 998244353;
 using base = modint<mod>;
 
 void solve() {
